Исправление

+++ src/main/java/jm/task/core/jdbc/dao/UserDaoHibernateImpl.java
@@ -2,6 +2,7 @@ package jm.task.core.jdbc.dao;
 
 import jm.task.core.jdbc.model.User;
 import jm.task.core.jdbc.util.Util;
+import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
 
@@ -22,6 +23,8 @@ public class UserDaoHibernateImpl implements UserDao {
             String sql = "CREATE TABLE IF NOT EXISTS users (id INT PRIMARY KEY AUTO_INCREMENT, name VARCHAR(20), lastName VARCHAR(20), age INT)";
             session.createSQLQuery(sql).executeUpdate();
             transaction.commit();
+        } catch (HibernateException e) {
+            e.printStackTrace();
         }
     }
 
@@ -32,6 +35,8 @@ public class UserDaoHibernateImpl implements UserDao {
             String sql = "DROP TABLE IF EXISTS users";
             session.createSQLQuery(sql).executeUpdate();
             transaction.commit();
+        } catch (HibernateException e) {
+            e.printStackTrace();
         }
     }
 
@@ -52,7 +57,7 @@ public class UserDaoHibernateImpl implements UserDao {
 
             transaction.commit();
 
-        } catch (Exception e) {
+        } catch (HibernateException e) {
             if (transaction != null) {
                 transaction.rollback();
             }
@@ -62,12 +67,18 @@ public class UserDaoHibernateImpl implements UserDao {
 
     @Override
     public void removeUserById(long id) {
-        Transaction transaction;
         try (Session session = Util.getSessionFactory().openSession()) {
-            transaction = session.beginTransaction();
+            Transaction transaction = session.beginTransaction();
             User user = session.get(User.class, id);
-            session.delete(user);
-            transaction.commit();
+
+            if (user != null) {
+                session.delete(user);
+                transaction.commit();
+            } else {
+                transaction.rollback();
+            }
+        } catch (HibernateException e) {
+            e.printStackTrace();
         }
     }
 
@@ -75,16 +86,25 @@ public class UserDaoHibernateImpl implements UserDao {
     public List<User> getAllUsers() {
         try (Session session = Util.getSessionFactory().openSession()) {
             return session.createQuery("from User", User.class).list();
+        } catch (HibernateException e) {
+            e.printStackTrace();
+            return null;
         }
     }
 
     @Override
     public void cleanUsersTable() {
+        Transaction transaction = null;
         try (Session session = Util.getSessionFactory().openSession()) {
-            Transaction transaction = session.beginTransaction();
+            transaction = session.beginTransaction();
             String sql = "TRUNCATE TABLE users";
             session.createSQLQuery(sql).executeUpdate();
             transaction.commit();
+        } catch (HibernateException e) {
+            if (transaction != null) {
+                transaction.rollback();
+            }
+            e.printStackTrace();
         }
     }
 }
